extract PatientService

+++ src/main/kotlin/dev/mtpeter/rsqrecruitmenttask/patient/PatientRouter.kt
@@ -4,9 +4,11 @@ import dev.mtpeter.rsqrecruitmenttask.configuration.TenantAwareRouting
 import dev.mtpeter.rsqrecruitmenttask.visit.VisitRepository
 import kotlinx.coroutines.async
 import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.toList
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
+import org.springframework.data.domain.Page
 import org.springframework.data.domain.PageImpl
 import org.springframework.data.domain.PageRequest
 import org.springframework.data.domain.Sort
@@ -36,38 +38,33 @@ class PatientRouter {
 @Transactional
 @Component
 class PatientHandler(
-    private val patientRepository: PatientRepository,
-    private val visitRepository: VisitRepository
+    private val patientService: PatientService
 ) {
 
     suspend fun getAllPatientsPaged(request: ServerRequest): ServerResponse = coroutineScope {
         val pageNo = request.queryParamOrNull("page")?.toIntOrNull() ?: 0
         val pageSize = request.queryParamOrNull("size")?.toIntOrNull() ?: 10
         val sort = Sort.by("lastName").and(Sort.by("firstName"))
-        val pageRequest = PageRequest.of(pageNo, pageSize, sort)
 
-        val patients = async { patientRepository.findAllBy(pageRequest).toList() }
-        val total = async { patientRepository.count() }
-
-        val page = PageImpl(patients.await(), pageRequest, total.await())
+        val page = patientService.getPagedPatients(pageNo, pageSize, sort)
         ServerResponse.ok().bodyValueAndAwait(page)
     }
 
     suspend fun getAllPatients(request: ServerRequest): ServerResponse {
-        val patientFlow = patientRepository.findAll()
+        val patientFlow = patientService.getAllPatients()
         return ServerResponse.ok().bodyAndAwait(patientFlow)
     }
 
     suspend fun getPatientById(request: ServerRequest): ServerResponse {
         val id = request.pathVariable("id").toLongOrNull() ?: return ServerResponse.badRequest().buildAndAwait()
-        val patient = patientRepository.findById(id) ?: return ServerResponse.notFound().buildAndAwait()
 
+        val patient = patientService.getPatientById(id) ?: return ServerResponse.notFound().buildAndAwait()
         return ServerResponse.ok().bodyValueAndAwait(patient)
     }
 
     suspend fun saveNewPatient(request: ServerRequest): ServerResponse {
         val patientDTO = request.awaitBody<PatientDTO>()
-        val saved = patientRepository.save(patientDTO.toPatient())
+        val saved = patientService.saveNewPatient(patientDTO)
 
         return ServerResponse.created(request.uriBuilder().path("/${saved.id}").build())
             .bodyValueAndAwait(saved)
@@ -77,10 +74,7 @@ class PatientHandler(
         val id = request.pathVariable("id").toLongOrNull() ?: return ServerResponse.badRequest().buildAndAwait()
         val patientDTO = request.awaitBody<PatientDTO>()
 
-        val exists = patientRepository.existsById(id)
-        if(!exists) return ServerResponse.notFound().buildAndAwait()
-
-        val saved = patientRepository.save(patientDTO.toPatient(id))
+        val saved = patientService.updatePatient(id, patientDTO) ?: return ServerResponse.notFound().buildAndAwait()
         return ServerResponse.ok().bodyValueAndAwait(saved)
     }
 
@@ -88,14 +82,54 @@ class PatientHandler(
         val id = request.pathVariable("id").toLongOrNull() ?: return ServerResponse.badRequest().buildAndAwait()
         val cascade = request.queryParamOrNull("cascade")?.toBoolean() ?: false
 
-        if(!patientRepository.existsById(id))
-            return ServerResponse.notFound().buildAndAwait()
-        if(!cascade && visitRepository.existsByPatientId(id))
-            return ServerResponse.status(HttpStatus.CONFLICT).buildAndAwait()
-        if(cascade)
-            visitRepository.removeByPatientId(id)
-
-        patientRepository.deleteById(id)
-        return ServerResponse.noContent().buildAndAwait()
+        return when(patientService.deletePatient(id, cascade)) {
+            is PatientNotFound -> ServerResponse.notFound().buildAndAwait()
+            is PatientHasVisits -> ServerResponse.status(HttpStatus.CONFLICT).buildAndAwait()
+            is DeletedSuccess -> ServerResponse.noContent().buildAndAwait()
+        }
     }
 }
+
+@Component
+class PatientService(
+    private val patientRepository: PatientRepository,
+    private val visitRepository: VisitRepository
+) {
+
+    fun getAllPatients(): Flow<Patient> = patientRepository.findAll()
+
+    suspend fun getPatientById(id: Long) = patientRepository.findById(id)
+
+    @Transactional
+    suspend fun getPagedPatients(pageNo: Int, pageSize: Int, sort: Sort): Page<Patient> = coroutineScope {
+        val pageRequest = PageRequest.of(pageNo, pageSize, sort)
+
+        val patients = async { patientRepository.findAllBy(pageRequest).toList() }
+        val total = async { patientRepository.count() }
+
+        PageImpl(patients.await(), pageRequest, total.await())
+    }
+
+    suspend fun saveNewPatient(patientDTO: PatientDTO): Patient = patientRepository.save(patientDTO.toPatient())
+
+    @Transactional
+    suspend fun updatePatient(id: Long, patientDTO: PatientDTO): Patient? {
+        if(!patientRepository.existsById(id)) return null
+        return patientRepository.save(patientDTO.toPatient(id))
+    }
+
+    @Transactional
+    suspend fun deletePatient(id: Long, cascade: Boolean): RemovalResult {
+        if (!patientRepository.existsById(id)) return PatientNotFound
+        if (!cascade && visitRepository.existsByPatientId(id)) return PatientHasVisits
+        if (cascade) visitRepository.removeByPatientId(id)
+
+        patientRepository.deleteById(id)
+        return DeletedSuccess
+    }
+}
+
+sealed interface RemovalResult
+object PatientNotFound : RemovalResult
+object PatientHasVisits : RemovalResult
+object DeletedSuccess : RemovalResult
